+++ utils/dynstream/batcher.go
@@ -14,11 +14,11 @@
 package dynstream
 
 type batchConfig struct {
-	batchCount int
-	batchBytes int
+	batchCount uint64
+	batchBytes uint64
 }
 
-func newBatchConfig(batchCount int, batchBytes int) batchConfig {
+func newBatchConfig(batchCount uint64, batchBytes uint64) batchConfig {
 	if batchCount <= 0 {
 		batchCount = 1
 	}
@@ -30,33 +30,33 @@ func newBatchConfig(batchCount int, batchBytes int) batchConfig {
 
 type batcher[T any] struct {
 	config   batchConfig
-	maxCount int
+	maxCount uint64
 
 	buf        []T
-	totalBytes int
+	totalBytes uint64
 }
 
-func newBatcher[T any](config batchConfig, initialCap int) batcher[T] {
+func newBatcher[T any](config batchConfig, initialCap uint64) batcher[T] {
 	if initialCap < 0 {
 		initialCap = 0
 	}
 	return batcher[T]{config: config, buf: make([]T, 0, initialCap)}
 }
 
-func (b *batcher[T]) setLimit(config batchConfig, maxCount int) {
+func (b *batcher[T]) setLimit(config batchConfig, maxCount uint64) {
 	b.config = config
 	b.maxCount = maxCount
 	b.totalBytes = 0
 	b.buf = b.buf[:0]
 }
 
-func (b *batcher[T]) addEvent(event T, eventBytes int) {
+func (b *batcher[T]) addEvent(event T, eventBytes uint64) {
 	b.buf = append(b.buf, event)
 	b.totalBytes += eventBytes
 }
 
 func (b *batcher[T]) isFull() bool {
-	if b.maxCount > 0 && len(b.buf) >= b.maxCount {
+	if b.maxCount > 0 && uint64(len(b.buf)) >= b.maxCount {
 		return true
 	}
 	if b.config.batchBytes > 0 && b.totalBytes >= b.config.batchBytes {
@@ -65,7 +65,7 @@ func (b *batcher[T]) isFull() bool {
 	return false
 }
 
-func (b *batcher[T]) flush() ([]T, int) {
+func (b *batcher[T]) flush() ([]T, uint64) {
 	return b.buf, b.totalBytes
 }
